+++ CumulusMX/Stations/EcowittCloudStation.cs
@@ -17,6 +17,10 @@ namespace CumulusMX
 		private Version deviceFirmware;
 		private int lastHour = -1;
 
+		private int PrimaryTHSensor;
+		private int PrimaryIndoorTHSensor;
+		private int[] MapWN34 = new int[9];
+
 		public EcowittCloudStation(Cumulus cumulus, WeatherStation station = null) : base(cumulus, station != null)
 		{
 			this.station = station;
@@ -36,6 +40,26 @@ namespace CumulusMX
 			{
 				cumulus.LogErrorMessage("API.GetHistoricData: Missing Ecowitt API data in the configuration, aborting!");
 				return;
+			}
+
+			// sensor mappings
+			if (mainStation)
+			{
+				PrimaryTHSensor = cumulus.Gw1000PrimaryTHSensor;
+				PrimaryIndoorTHSensor = cumulus.Gw1000PrimaryIndoorTHSensor;
+				for (int i = 0; i < cumulus.EcowittMapWN34.Length; i++)
+				{
+					MapWN34[i] = cumulus.EcowittMapWN34[i];
+				}
+			}
+			else
+			{
+				PrimaryTHSensor = cumulus.ExtraPrimaryTHSensor;
+				PrimaryIndoorTHSensor = cumulus.ExtraPrimaryIndoorTHSensor;
+				for (int i = 0; i < cumulus.ExtraMapWN34.Length; i++)
+				{
+					MapWN34[i] = cumulus.ExtraMapWN34[i];
+				}
 			}
 
 
@@ -61,12 +85,12 @@ namespace CumulusMX
 				// does not provide pressure trend strings
 				cumulus.StationOptions.UseCumulusPresstrendstr = true;
 
-				if (cumulus.Gw1000PrimaryTHSensor == 0)
+				if (PrimaryTHSensor == 0)
 				{
 					// We are using the primary T/H sensor
 					cumulus.LogMessage("Using the default outdoor temp/hum sensor data");
 				}
-				else if (cumulus.Gw1000PrimaryTHSensor == 99)
+				else if (PrimaryTHSensor == 99)
 				{
 					cumulus.LogMessage("Overriding the default outdoor temp/hum data with the internal sensor");
 					cumulus.StationOptions.CalculatedDP = true;
@@ -74,7 +98,7 @@ namespace CumulusMX
 				else
 				{
 					// We are not using the primary T/H sensor
-					cumulus.LogMessage("Overriding the default outdoor temp/hum data with Extra temp/hum sensor #" + cumulus.Gw1000PrimaryTHSensor);
+					cumulus.LogMessage("Overriding the default outdoor temp/hum data with Extra temp/hum sensor #" + PrimaryTHSensor);
 				}
 
 				if (cumulus.Gw1000PrimaryRainSensor == 0)
@@ -87,7 +111,7 @@ namespace CumulusMX
 					cumulus.LogMessage("Using the piezo rain sensor data");
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 0)
+				if (PrimaryIndoorTHSensor == 0)
 				{
 					// We are using the primary indoor T/H sensor
 					cumulus.LogMessage("Using the default indoor temp/hum sensor data");
@@ -95,7 +119,7 @@ namespace CumulusMX
 				else
 				{
 					// We are not using the primary indoor T/H sensor
-					cumulus.LogMessage("Overriding the default indoor temp/hum data with Extra temp/hum sensor #" + cumulus.Gw1000PrimaryIndoorTHSensor);
+					cumulus.LogMessage("Overriding the default indoor temp/hum data with Extra temp/hum sensor #" + PrimaryIndoorTHSensor);
 				}
 
 				DataTimeoutMins = cumulus.EcowittCloudDataUpdateInterval + 2;
@@ -114,7 +138,7 @@ namespace CumulusMX
 				cumulus.Units.LeafWetnessUnitText = "%";
 			}
 
-			ecowittApi = new EcowittApi(cumulus, this);
+			ecowittApi = new EcowittApi(cumulus, this, mainStation);
 
 			// Only perform the Start-up if we are a proper station, not a Extra Sensor
 			try
@@ -343,7 +367,7 @@ namespace CumulusMX
 				if (mainStation)
 				{
 					// Outdoor temp/hum
-					if (cumulus.Gw1000PrimaryTHSensor == 0)
+					if (PrimaryTHSensor == 0)
 					{
 						if (data.outdoor == null)
 						{
@@ -379,7 +403,7 @@ namespace CumulusMX
 						try
 						{
 							// user has mapped the indoor sensor to the outdoor sensor
-							if (cumulus.Gw1000PrimaryTHSensor == 99)
+							if (PrimaryTHSensor == 99)
 							{
 								var time = Utils.FromUnixTime(data.outdoor.temperature.time);
 								DoOutdoorTemp(data.indoor.temperature.value, time);
@@ -391,7 +415,7 @@ namespace CumulusMX
 								DoCloudBaseHeatIndex(time);
 							}
 
-							if (cumulus.Gw1000PrimaryIndoorTHSensor == 0)
+							if (PrimaryIndoorTHSensor == 0)
 							{
 								DoIndoorTemp(data.indoor.temperature.value);
 								DoIndoorHumidity(data.indoor.humidity.value);
@@ -673,12 +697,12 @@ namespace CumulusMX
 		{
 			if (data.temp_and_humidity_ch1 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 1)
+				if (PrimaryTHSensor == 1)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch1.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch1.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 1)
+				if (PrimaryIndoorTHSensor == 1)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch1.temperature.value);
 				}
@@ -688,12 +712,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch1.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 1)
+					if (PrimaryTHSensor == 1)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch1.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch1.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 1)
+					if (PrimaryIndoorTHSensor == 1)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch1.humidity.value);
 					}
@@ -710,12 +734,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch2 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 2)
+				if (PrimaryTHSensor == 2)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch2.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch2.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 2)
+				if (PrimaryIndoorTHSensor == 2)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch2.temperature.value);
 				}
@@ -725,12 +749,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch2.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 2)
+					if (PrimaryTHSensor == 2)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch2.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch2.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 2)
+					if (PrimaryIndoorTHSensor == 2)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch2.humidity.value);
 					}
@@ -747,12 +771,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch3 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 3)
+				if (PrimaryTHSensor == 3)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch3.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch3.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 3)
+				if (PrimaryIndoorTHSensor == 3)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch3.temperature.value);
 				}
@@ -762,12 +786,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch3.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 3)
+					if (PrimaryTHSensor == 3)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch3.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch3.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 3)
+					if (PrimaryIndoorTHSensor == 3)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch3.humidity.value);
 					}
@@ -784,12 +808,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch4 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 4)
+				if (PrimaryTHSensor == 4)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch4.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch4.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 4)
+				if (PrimaryIndoorTHSensor == 4)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch4.temperature.value);
 				}
@@ -799,12 +823,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch4.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 4)
+					if (PrimaryTHSensor == 4)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch4.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch4.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 4)
+					if (PrimaryIndoorTHSensor == 4)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch4.humidity.value);
 					}
@@ -821,12 +845,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch5 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 5)
+				if (PrimaryTHSensor == 5)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch5.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch5.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 5)
+				if (PrimaryIndoorTHSensor == 5)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch5.temperature.value);
 				}
@@ -836,12 +860,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch5.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 5)
+					if (PrimaryTHSensor == 5)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch5.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch5.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 5)
+					if (PrimaryIndoorTHSensor == 5)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch5.humidity.value);
 					}
@@ -858,12 +882,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch6 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 6)
+				if (PrimaryTHSensor == 6)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch6.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch6.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 6)
+				if (PrimaryIndoorTHSensor == 6)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch6.temperature.value);
 				}
@@ -873,12 +897,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch6.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 6)
+					if (PrimaryTHSensor == 6)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch6.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch6.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 6)
+					if (PrimaryIndoorTHSensor == 6)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch6.humidity.value);
 					}
@@ -895,12 +919,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch7 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 7)
+				if (PrimaryTHSensor == 7)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch7.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch7.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 7)
+				if (PrimaryIndoorTHSensor == 7)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch7.temperature.value);
 				}
@@ -910,12 +934,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch7.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 7)
+					if (PrimaryTHSensor == 7)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch7.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch7.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 7)
+					if (PrimaryIndoorTHSensor == 7)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch7.humidity.value);
 					}
@@ -932,12 +956,12 @@ namespace CumulusMX
 
 			if (data.temp_and_humidity_ch8 != null)
 			{
-				if (cumulus.Gw1000PrimaryTHSensor == 8)
+				if (PrimaryTHSensor == 8)
 				{
 					station.DoOutdoorTemp(data.temp_and_humidity_ch8.temperature.value, Utils.FromUnixTime(data.temp_and_humidity_ch8.temperature.time));
 				}
 
-				if (cumulus.Gw1000PrimaryIndoorTHSensor == 8)
+				if (PrimaryIndoorTHSensor == 8)
 				{
 					station.DoIndoorTemp(data.temp_and_humidity_ch8.temperature.value);
 				}
@@ -947,12 +971,12 @@ namespace CumulusMX
 				// Not all sensor types have humidity
 				if (data.temp_and_humidity_ch8.humidity != null)
 				{
-					if (cumulus.Gw1000PrimaryTHSensor == 8)
+					if (PrimaryTHSensor == 8)
 					{
 						station.DoOutdoorHumidity(data.temp_and_humidity_ch8.humidity.value, Utils.FromUnixTime(data.temp_and_humidity_ch8.humidity.time));
 					}
 
-					if (cumulus.Gw1000PrimaryIndoorTHSensor == 8)
+					if (PrimaryIndoorTHSensor == 8)
 					{
 						station.DoIndoorHumidity(data.temp_and_humidity_ch8.humidity.value);
 					}
@@ -972,91 +996,91 @@ namespace CumulusMX
 		{
 			if (data.temp_ch1 != null)
 			{
-				if (cumulus.EcowittMapWN34[1] == 0)
+				if (MapWN34[1] == 0)
 				{
 					station.DoUserTemp(data.temp_ch1.temperature.value, 1);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch1.temperature.value, cumulus.EcowittMapWN34[1]);
+					station.DoSoilTemp(data.temp_ch1.temperature.value, MapWN34[1]);
 				}
 			}
 
 			if (data.temp_ch2 != null)
 			{
-				if (cumulus.EcowittMapWN34[2] == 0)
+				if (MapWN34[2] == 0)
 				{
 					station.DoUserTemp(data.temp_ch2.temperature.value, 2);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch2.temperature.value, cumulus.EcowittMapWN34[2]);
+					station.DoSoilTemp(data.temp_ch2.temperature.value, MapWN34[2]);
 				}
 			}
 
 			if (data.temp_ch3 != null)
 			{
-				if (cumulus.EcowittMapWN34[3] == 0)
+				if (MapWN34[3] == 0)
 				{
 					station.DoUserTemp(data.temp_ch3.temperature.value, 3);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch3.temperature.value, cumulus.EcowittMapWN34[3]);
+					station.DoSoilTemp(data.temp_ch3.temperature.value, MapWN34[3]);
 				}
 			}
 
 			if (data.temp_ch4 != null)
 			{
-				if (cumulus.EcowittMapWN34[4] == 0)
+				if (MapWN34[4] == 0)
 				{
 					station.DoUserTemp(data.temp_ch4.temperature.value, 4);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch4.temperature.value, cumulus.EcowittMapWN34[4]);
+					station.DoSoilTemp(data.temp_ch4.temperature.value, MapWN34[4]);
 				}
 			}
 
 			if (data.temp_ch5 != null)
 			{
-				if (cumulus.EcowittMapWN34[5] == 0)
+				if (MapWN34[5] == 0)
 				{
 					station.DoUserTemp(data.temp_ch5.temperature.value, 5);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch5.temperature.value, cumulus.EcowittMapWN34[5]);
+					station.DoSoilTemp(data.temp_ch5.temperature.value, MapWN34[5]);
 				}
 			}
 
 			if (data.temp_ch6 != null)
 			{
-				if (cumulus.EcowittMapWN34[6] == 0)
+				if (MapWN34[6] == 0)
 				{
 					station.DoUserTemp(data.temp_ch6.temperature.value, 6);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch6.temperature.value, cumulus.EcowittMapWN34[6]);
+					station.DoSoilTemp(data.temp_ch6.temperature.value, MapWN34[6]);
 				}
 			}
 
 			if (data.temp_ch7 != null)
 			{
-				if (cumulus.EcowittMapWN34[7] == 0)
+				if (MapWN34[7] == 0)
 				{
 					station.DoUserTemp(data.temp_ch7.temperature.value, 7);
 				}
 				else
 				{
-					station.DoSoilTemp(data.temp_ch7.temperature.value, cumulus.EcowittMapWN34[7]);
+					station.DoSoilTemp(data.temp_ch7.temperature.value, MapWN34[7]);
 				}
 			}
 
 			if (data.temp_ch8 != null)
 			{
-				if (cumulus.EcowittMapWN34[8] == 0)
+				if (MapWN34[8] == 0)
 				{
 					station.DoUserTemp(data.temp_ch8.temperature.value, 8);
 				}